+++ bridgetown-core/lib/bridgetown-core/commands/serve.rb
@@ -122,7 +122,7 @@ module Bridgetown
 
       def start_up_webrick(opts, destination)
         @server = WEBrick::HTTPServer.new(webrick_opts(opts)).tap { |o| o.unmount("") }
-        @server.mount(opts["baseurl"].to_s, Servlet, destination, file_handler_opts)
+        @server.mount(opts["base_path"].to_s, Servlet, destination, file_handler_opts)
 
         Bridgetown.logger.info "Server address:", server_address(@server, opts)
         launch_browser @server, opts if opts["open_url"]