+++ bridgetown-core/lib/bridgetown-core/concerns/site/configurable.rb
@@ -34,6 +34,22 @@ class Bridgetown::Site
       config[:content_engine] == "resource"
     end
 
+    # Returns a base path from which the site is served (aka `/cool-site`) or
+    # `/` if served from root.
+    #
+    # @param strip_slash_only [Boolean] set to true if you wish "/" to be returned as ""
+    # @return [String]
+    def base_path(strip_slash_only: false)
+      (config[:base_path] || config[:baseurl]).yield_self do |path|
+        strip_slash_only ? path.to_s.sub(%r{^/$}, "") : path
+      end
+    end
+
+    def baseurl
+      Bridgetown::Deprecator.deprecation_message "Site#baseurl is now Site#base_path"
+      base_path(strip_slash_only: true).presence
+    end
+
     def defaults_reader
       @defaults_reader ||= Bridgetown::DefaultsReader.new(self)
     end