+++ bridgetown-core/test/test_webpack_command.rb
@@ -13,7 +13,7 @@ class TestWebpackCommand < BridgetownUnitTest
 
   context "the webpack command" do
     setup do
-      @path = "new-site"
+      @path = SecureRandom.alphanumeric
       @full_path = File.expand_path(@path, Dir.pwd)
 
       capture_stdout { Bridgetown::Commands::Base.start(["new", @path]) }
@@ -63,18 +63,14 @@ class TestWebpackCommand < BridgetownUnitTest
       refute_file_contains %r!OLD_VERSION!, webpack_defaults
     end
 
-    # TODO: Figure out how to make this test pass
-    # Info: https://github.com/bridgetownrb/bridgetown/pull/270#issuecomment-841709898
+    should "enable postcss in webpack config" do
+      refute_file_contains %r!mode: 'postcss'!, webpack_defaults
 
-    # should "enable postcss in webpack config" do
-    #   webpack_defaults = File.join(@full_path, "webpack.defaults.js")
-    #   refute_file_contains %r!postcss-loader!, webpack_defaults
+      @cmd.inside(@full_path) do
+        capture_stdout { @cmd.invoke(:webpack, ["enable-postcss"]) }
+      end
 
-    #   @cmd.inside(@full_path) do
-    #     capture_stdout { @cmd.invoke(:webpack, ["enable-postcss"]) }
-    #   end
-
-    #   assert_file_contains %r!postcss-loader!, webpack_defaults
-    # end
+      assert_file_contains %r!mode: 'postcss'!, webpack_defaults
+    end
   end
 end